refactor(accordion): split accordion into modules
Split 838-line accordion/mod.rs into focused modules:
- mod.rs: struct Accordion + builders + state methods + tests (~310 lines)
- render.rs: impl View for Accordion (~130 lines)
- types.rs: AccordionSection struct + builders (~80 lines)

All 28 accordion tests continue to pass.
